fix compound command handling
Not all programm messages with several program message units are compound statements. Currently the parser
uses hints like the colon or asterisk at the beginning of the next command to determine whether or not the
next program message unit should be treated as compound command.

The aproach failes for sequences of valid device specific commands unless they start with colons. Consider:

TREE:TESTA?; TEXT?

In this case , second unit is treated as compound command "TREE:TEXT?" which would not exist.

Therefore modify the parser to first check if the unit is a valid program message as it stands, before
handling it as compound command